pyproject.toml
[project]
name = "glukhov-sales-engine"
version = "0.1.0"
description = "Automated lead collection system for GlukhovSystems"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "telethon>=1.36.0",
    "playwright>=1.45.0",
    "rapidfuzz>=3.9.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "redis>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.100.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "testcontainers>=4.0.0",
    "respx>=0.21.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
